+++ spec/rubocop/cop/naming/rescued_exceptions_variable_name_spec.rb
@@ -113,6 +113,100 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           RUBY
         end
       end
+
+      context 'with lower letters class name' do
+        it 'does not register an offense' do
+          expect_no_offenses(<<~RUBY)
+            begin
+              something
+            rescue my_exception
+              # do something
+            end
+          RUBY
+        end
+      end
+
+      context 'with method as `Exception`' do
+        it 'does not register an offense without variable name' do
+          expect_no_offenses(<<~RUBY)
+            begin
+              something
+            rescue ActiveSupport::JSON.my_method
+              # do something
+            end
+          RUBY
+        end
+
+        it 'does not register an offense with expected variable name' do
+          expect_no_offenses(<<~RUBY)
+            begin
+              something
+            rescue ActiveSupport::JSON.my_method => e
+              # do something
+            end
+          RUBY
+        end
+
+        it 'registers an offense with unexpected variable name' do
+          expect_offense(<<~RUBY)
+            begin
+              something
+            rescue ActiveSupport::JSON.my_method => exc
+                                                    ^^^ Use `e` instead of `exc`.
+              # do something
+            end
+          RUBY
+
+          expect_correction(<<~RUBY)
+            begin
+              something
+            rescue ActiveSupport::JSON.my_method => e
+              # do something
+            end
+          RUBY
+        end
+      end
+
+      context 'with splat operator as `Exception` list' do
+        it 'does not register an offense without variable name' do
+          expect_no_offenses(<<~RUBY)
+            begin
+              something
+            rescue *handled
+              # do something
+            end
+          RUBY
+        end
+
+        it 'does not register an offense with expected variable name' do
+          expect_no_offenses(<<~RUBY)
+            begin
+              something
+            rescue *handled => e
+              # do something
+            end
+          RUBY
+        end
+
+        it 'registers an offense with unexpected variable name' do
+          expect_offense(<<~RUBY)
+            begin
+              something
+            rescue *handled => exc
+                               ^^^ Use `e` instead of `exc`.
+              # do something
+            end
+          RUBY
+
+          expect_correction(<<~RUBY)
+            begin
+              something
+            rescue *handled => e
+              # do something
+            end
+          RUBY
+        end
+      end
     end
 
     context 'with implicit rescue' do
@@ -121,8 +215,8 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_offense(<<~RUBY)
             begin
               something
-            rescue exc
-                   ^^^ Use `e` instead of `exc`.
+            rescue => exc
+                      ^^^ Use `e` instead of `exc`.
               # do something
             end
           RUBY
@@ -130,7 +224,7 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_correction(<<~RUBY)
             begin
               something
-            rescue e
+            rescue => e
               # do something
             end
           RUBY
@@ -140,8 +234,8 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_offense(<<~RUBY)
             begin
               something
-            rescue _exc
-                   ^^^^ Use `_e` instead of `_exc`.
+            rescue => _exc
+                      ^^^^ Use `_e` instead of `_exc`.
               # do something
             end
           RUBY
@@ -149,7 +243,7 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_correction(<<~RUBY)
             begin
               something
-            rescue _e
+            rescue => _e
               # do something
             end
           RUBY
@@ -159,7 +253,7 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_no_offenses(<<~RUBY)
             begin
               something
-            rescue e
+            rescue => e
               # do something
             end
           RUBY
@@ -169,7 +263,7 @@ RSpec.describe RuboCop::Cop::Naming::RescuedExceptionsVariableName, :config do
           expect_no_offenses(<<~RUBY)
             begin
               something
-            rescue _e
+            rescue => _e
               # do something
             end
           RUBY